+++ lib/Models/WebMapServiceCatalogGroup.ts
@@ -273,6 +273,13 @@ class GetCapabilitiesStratum extends LoadableStratum(
       "hideLegendInWorkbench",
       this.catalogGroup.hideLegendInWorkbench
     );
+
+    if (this.catalogGroup.itemProperties !== undefined) {
+      Object.keys(this.catalogGroup.itemProperties).map((k: any) => {
+        if (this.catalogGroup.itemProperties !== undefined)
+          model.setTrait(stratum, k, this.catalogGroup.itemProperties[k]);
+      });
+    }
     model.createGetCapabilitiesStratumFromParent(this.capabilities);
   }